requirements.txt
streamlit==1.52.2
opencv-python==4.12.0.88
mediapipe==0.10.9
face-recognition==1.3.0
pyzbar==0.1.9
psycopg2-binary==2.9.11
pandas==2.3.3
numpy==2.2.6
